Add church bools
