+++ tests/factorial_integration_test.rs
@@ -5,7 +5,7 @@ use cauliflambda::parse_formula;
 fn factorial_of_four_can_be_computed() {
     let formula = parse_formula(include_str!("factorial.lc")).unwrap();
     assert!(formula.diagnostics.0.is_empty());
-    let last_step = evaluate(formula.value, Options::default().add_church_numerals())
+    let last_step = evaluate(formula.value, Options::default().add_church_encoding())
         .take(100_000)
         .last()
         .unwrap();
